refactor: 비속어 검증을 일반 기능으로 분류
